[rename] collection state store and add a service module to be used by UI components
